Change: Move useUserSessionTimeout hook to hooks directory
