+++ src/web/entities/BulkTags.jsx
@@ -14,7 +14,7 @@ import PropTypes from 'web/utils/proptypes';
 import SelectionType from 'web/utils/selectiontype';
 
 import useGmp from 'web/utils/useGmp';
-import useUserSessionTimeout from 'web/utils/useUserSessionTimeout';
+import useUserSessionTimeout from 'web/hooks/useUserSessionTimeout';
 
 import TagDialog from 'web/pages/tags/dialog';
