+++ src/web/hooks/__tests__/useUserSessionTimeout.jsx
@@ -10,7 +10,7 @@ import date from 'gmp/models/date';
 
 import {setSessionTimeout as setSessionTimeoutAction} from 'web/store/usersettings/actions';
 
-import {rendererWith, fireEvent} from '../testing';
+import {rendererWith, fireEvent} from 'web/utils/testing';
 
 import useUserSessionTimeout from '../useUserSessionTimeout';
