+++ src/web/pages/portlists/listpage.jsx
@@ -12,7 +12,7 @@ import {PORTLISTS_FILTER_FILTER} from 'gmp/models/filter';
 import {isDefined, hasValue} from 'gmp/utils/identity';
 
 import useCapabilities from 'web/hooks/useCapabilities';
-import useUserSessionTimeout from 'web/utils/useUserSessionTimeout';
+import useUserSessionTimeout from 'web/hooks/useUserSessionTimeout';
 
 import useGmp from 'web/utils/useGmp';
 import usePageFilter from 'web/hooks/usePageFilter';
